# src/PirateShipCollection/wwwroot/swagger-original.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Swagger file for the SWE 2021 course.",
    "version" : "1.0.0",
    "title" : "Swashbuckling Buccaneer Pirate Store"
  },
  "host" : "shipcollection.swagger.io",
  "basePath" : "/v2",
  "tags" : [ {
    "name" : "ship",
    "description" : "Access to the swashbuckling buccaneers pirate store."
  } ],
  "schemes" : [ "http" ],
  "paths" : {
    "/ship/{shipId}" : {
      "get" : {
        "tags" : [ "ship" ],
        "summary" : "Find ship by ID",
        "description" : "Returns a single ship",
        "operationId" : "getShipById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "shipId",
          "in" : "path",
          "description" : "ID of ship to return",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Ship"
            }
          },
          "400" : {
            "description" : "Invalid ID supplied"
          },
          "404" : {
            "description" : "Ship not found"
          }
        }
      },
      "delete" : {
        "tags" : [ "ship" ],
        "summary" : "Deletes a pet",
        "description" : "",
        "operationId" : "deletePet",
        "produces" : [ "application/xml", "application/json" ],
        "parameters" : [ {
          "name" : "shipId",
          "in" : "path",
          "description" : "Ship id to delete",
          "required" : true,
          "type" : "integer",
          "format" : "uint64"
        } ],
        "responses" : {
          "400" : {
            "description" : "Invalid ID supplied"
          },
          "404" : {
            "description" : "Ship not found"
          }
        }
      }
    },
    "/ship" : {
      "post" : {
        "tags" : [ "ship" ],
        "summary" : "Add a new ship to the store",
        "description" : "",
        "operationId" : "addShip",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "Ship object that needs to be added to the collection",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Ship"
          }
        } ],
        "responses" : {
          "405" : {
            "description" : "Invalid input"
          }
        }
      },
      "put" : {
        "tags" : [ "ship" ],
        "summary" : "Update an existing ship",
        "description" : "",
        "operationId" : "updateShip",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "Ship object that needs to be added to the store",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Ship"
          }
        } ],
        "responses" : {
          "400" : {
            "description" : "Invalid ID supplied"
          },
          "404" : {
            "description" : "Ship not found"
          },
          "405" : {
            "description" : "Validation exception"
          }
        }
      }
    }
  },
  "definitions" : {
    "Ship" : {
      "type" : "object",
      "required" : [ "name" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "category" : {
          "type" : "string"
        },
        "name" : {
          "type" : "string",
          "example" : "Black Pearl"
        },
        "length" : {
          "type" : "integer",
          "format" : "uint64"
        },
        "width" : {
          "type" : "integer",
          "format" : "uint64"
        },
        "height" : {
          "type" : "integer",
          "format" : "uint64"
        },
        "weight" : {
          "type" : "integer",
          "format" : "uint64"
        }
      },
      "example" : {
        "name" : "Black Pearl",
        "length" : 6,
        "width" : 1,
        "weight" : 5,
        "id" : 0,
        "category" : "category",
        "height" : 5
      }
    },
    "ApiResponse" : {
      "type" : "object",
      "properties" : {
        "code" : {
          "type" : "integer",
          "format" : "int32"
        },
        "type" : {
          "type" : "string"
        },
        "message" : {
          "type" : "string"
        }
      }
    }
  }
}